Create Task1&Task2.sol
Solidity tasks in which i print "hello solidity" and in second task printing state and local variables. State variables is like Global variables. It is stored on the blockchain. It costs some gas. It is declared at the contract level.
local variables declared inside function. Memory keyword stores on runtime.